fix casting issue #83 (long -> int32_t); + reindented;

+++ libraries/XMLWriter/XMLWriter.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: XMLWriter.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.7
+// VERSION: 0.1.8
 //    DATE: 2013-11-06
 // PURPOSE: Simple XML library
 //
@@ -14,6 +14,7 @@
 // 0.1.05  2015-05-23 added XMLWRITER_MAXTAGSIZE 15 (to support KML coordinates tag)
 // 0.1.6   2016-03-16 added incrIndent(), decrIndent(), indent(), raw();
 // 0.1.7   2017-07-26 added const where possible
+// 0.1.8   2017-12-09 fix casting issue #83 (long -> int32_t);
 //
 // Released to the public domain
 //
@@ -22,208 +23,208 @@
 
 XMLWriter::XMLWriter(Print* stream)
 {
-    _stream = stream;
-    reset();
+  _stream = stream;
+  reset();
 }
 
 void XMLWriter::reset()
 {
-    _indent = 0;
-    _indentStep = 2;
-    _idx = 0;
+  _indent = 0;
+  _indentStep = 2;
+  _idx = 0;
 }
 
 void XMLWriter::header()
 {
-    _stream->println(F("<?xml version=\"1.0\" encoding=\"UTF-8\"?>"));
+  _stream->println(F("<?xml version=\"1.0\" encoding=\"UTF-8\"?>"));
 }
 
 void XMLWriter::comment(const char* text, const bool multiLine)
 {
-    _stream->println();
-    if (!multiLine) indent();
-    _stream->print(F("<!-- "));
-    if (multiLine) _stream->println();
-    _stream->print(text);
-    if (multiLine) _stream->println();
-    _stream->println(F(" -->"));
+  _stream->println();
+  if (!multiLine) indent();
+  _stream->print(F("<!-- "));
+  if (multiLine) _stream->println();
+  _stream->print(text);
+  if (multiLine) _stream->println();
+  _stream->println(F(" -->"));
 }
 
 void XMLWriter::tagOpen(const char* tag, const bool newline)
 {
-    tagOpen(tag, "", newline);
+  tagOpen(tag, "", newline);
 }
 
 void XMLWriter::tagOpen(const char* tag, const char* name, const bool newline)
 {
-    // TODO STACK GUARD
-    strncpy(tagStack[_idx++], tag, XMLWRITER_MAXTAGSIZE);
-    tagStart(tag);
-    if (name[0] != 0) tagField("name", name);
-    tagEnd(newline, NOSLASH);
-    _indent += _indentStep;
+  // TODO STACK GUARD
+  strncpy(tagStack[_idx++], tag, XMLWRITER_MAXTAGSIZE);
+  tagStart(tag);
+  if (name[0] != 0) tagField("name", name);
+  tagEnd(newline, NOSLASH);
+  _indent += _indentStep;
 }
 
 void XMLWriter::tagClose(const bool ind)
 {
-    _indent -= _indentStep;
-    if (ind) indent();
-    _stream->print(F("</"));
-    _stream->print(tagStack[--_idx]);
-    _stream->println(F(">"));
+  _indent -= _indentStep;
+  if (ind) indent();
+  _stream->print(F("</"));
+  _stream->print(tagStack[--_idx]);
+  _stream->println(F(">"));
 }
 
 void XMLWriter::tagStart(const char *tag)
 {
-    indent();
-    _stream->print('<');
-    _stream->print(tag);
+  indent();
+  _stream->print('<');
+  _stream->print(tag);
 }
 
 void XMLWriter::tagField(const char *field, const char* str)
 {
-    _stream->print(' ');
-    _stream->print(field);
-    _stream->print(F("=\""));
+  _stream->print(' ');
+  _stream->print(field);
+  _stream->print(F("=\""));
 #ifdef XMLWRITER_ESCAPE_SUPPORT
-    escape(str);
+  escape(str);
 #else
-    _stream->print(str);
+  _stream->print(str);
 #endif
-    _stream->print('"');
+  _stream->print('"');
 }
 
 void XMLWriter::tagEnd(const bool newline, const bool addSlash)
 {
-    if (addSlash) _stream->print('/');
-    _stream->print('>');
-    if (newline) _stream->println();
+  if (addSlash) _stream->print('/');
+  _stream->print('>');
+  if (newline) _stream->println();
 }
 
 void XMLWriter::writeNode(const char* tag, const char* str)
 {
-    tagOpen(tag, "", NONEWLINE);
+  tagOpen(tag, "", NONEWLINE);
 #ifdef XMLWRITER_ESCAPE_SUPPORT
-    escape(str);
+  escape(str);
 #else
-    _stream->print(str);
+  _stream->print(str);
 #endif
-    tagClose(NOINDENT);
+  tagClose(NOINDENT);
 }
 
 void XMLWriter::setIndentSize(const uint8_t size)
 {
-    _indentStep = size;
+  _indentStep = size;
 }
 
 void XMLWriter::tagField(const char *field, const uint8_t value, const uint8_t base)
 {
-    tagField(field, (uint32_t) value, base);
+  tagField(field, (uint32_t) value, base);
 }
 
 void XMLWriter::tagField(const char *field, const uint16_t value, const uint8_t base)
 {
-    tagField(field, (uint32_t) value, base);
+  tagField(field, (uint32_t) value, base);
 }
 
 void XMLWriter::tagField(const char *field, const uint32_t value, const uint8_t base)
 {
-    _stream->print(' ');
-    _stream->print(field);
-    _stream->print(F("=\""));
-    _stream->print(value, base);
-    _stream->print('"');
+  _stream->print(' ');
+  _stream->print(field);
+  _stream->print(F("=\""));
+  _stream->print(value, base);
+  _stream->print('"');
 }
 
 void XMLWriter::tagField(const char *field, const int8_t value, const uint8_t base)
 {
-    tagField(field, (long) value, base);
+  tagField(field, (int32_t) value, base);
 }
 
 void XMLWriter::tagField(const char *field, const int16_t value, const uint8_t base)
 {
-    tagField(field, (long) value, base);
+  tagField(field, (int32_t) value, base);
 }
 
 void XMLWriter::tagField(const char *field, const int32_t value, const uint8_t base)
 {
-    _stream->print(' ');
-    _stream->print(field);
-    _stream->print(F("=\""));
-    _stream->print(value, base);
-    _stream->print('"');
+  _stream->print(' ');
+  _stream->print(field);
+  _stream->print(F("=\""));
+  _stream->print(value, base);
+  _stream->print('"');
 }
 
 
 void XMLWriter::tagField(const char *field, const bool value)
 {
-    _stream->print(' ');
-    _stream->print(field);
-    _stream->print(F("=\""));
-    _stream->print(value?F("true"):F("false"));
-    _stream->print('"');
+  _stream->print(' ');
+  _stream->print(field);
+  _stream->print(F("=\""));
+  _stream->print(value?F("true"):F("false"));
+  _stream->print('"');
 }
 
 void XMLWriter::tagField(const char *field, const double value, const uint8_t decimals)
 {
-    _stream->print(' ');
-    _stream->print(field);
-    _stream->print(F("=\""));
-    _stream->print(value, decimals);
-    _stream->print('"');
+  _stream->print(' ');
+  _stream->print(field);
+  _stream->print(F("=\""));
+  _stream->print(value, decimals);
+  _stream->print('"');
 }
 
 void XMLWriter::writeNode(const char* tag, const uint8_t value, const uint8_t base)
 {
-    writeNode(tag, (uint32_t) value, base);
+  writeNode(tag, (uint32_t) value, base);
 }
 
 void XMLWriter::writeNode(const char* tag, const uint16_t value, const uint8_t base)
 {
-    writeNode(tag, (uint32_t) value, base);
+  writeNode(tag, (uint32_t) value, base);
 }
 
 void XMLWriter::writeNode(const char* tag, const uint32_t value, const uint8_t base)
 {
-    tagOpen(tag, "", NONEWLINE);
-    _stream->print(value, base);  // todo: leading zero's
-    tagClose(NOINDENT);
+  tagOpen(tag, "", NONEWLINE);
+  _stream->print(value, base);  // todo: leading zero's
+  tagClose(NOINDENT);
 }
 
 void XMLWriter::writeNode(const char* tag, const int8_t value, const uint8_t base)
 {
-    writeNode(tag, (int32_t) value, base);
+  writeNode(tag, (int32_t) value, base);
 }
 
 void XMLWriter::writeNode(const char* tag, const int16_t value, const uint8_t base)
 {
-    writeNode(tag, (int32_t) value, base);
+  writeNode(tag, (int32_t) value, base);
 }
 
 void XMLWriter::writeNode(const char* tag, const int32_t value, const uint8_t base)
 {
-    tagOpen(tag, "", NONEWLINE);
-    _stream->print(value, base);  // todo: leading zero's
-    tagClose(NOINDENT);
+  tagOpen(tag, "", NONEWLINE);
+  _stream->print(value, base);  // todo: leading zero's
+  tagClose(NOINDENT);
 }
 
 void XMLWriter::writeNode(const char* tag, const bool value)
 {
-    tagOpen(tag, "", NONEWLINE);
-    _stream->print(value?F("true"):F("false"));
-    tagClose(NOINDENT);
+  tagOpen(tag, "", NONEWLINE);
+  _stream->print(value?F("true"):F("false"));
+  tagClose(NOINDENT);
 }
 
 void XMLWriter::writeNode(const char* tag, const double value, const uint8_t decimals)
 {
-    tagOpen(tag, "", NONEWLINE);
-    _stream->print(value, decimals);
-    tagClose(NOINDENT);
+  tagOpen(tag, "", NONEWLINE);
+  _stream->print(value, decimals);
+  tagClose(NOINDENT);
 }
 
 void XMLWriter::indent()
 {
-    for (uint8_t i=_indent; i>0; i--) _stream->print(' ');
+  for (uint8_t i=_indent; i>0; i--) _stream->print(' ');
 }
 
 ////////////////////////////////////////////////////////////////////
@@ -234,14 +235,14 @@ char expanded[][7] = { "&quot;", "&apos;","&lt;","&gt;","&amp;"}; // todo in fla
 
 void XMLWriter::escape(const char* str)
 {
-    char* p = (char *)str;
-    while(*p != 0)
-    {
-        char* q = strchr(c, *p);
-        if (q == NULL) _stream->print(*p);
-        else _stream->print(expanded[q - c]); // uint8_t idx = q-c;
-        p++;
-    }
+  char* p = (char *)str;
+  while(*p != 0)
+  {
+    char* q = strchr(c, *p);
+    if (q == NULL) _stream->print(*p);
+    else _stream->print(expanded[q - c]); // uint8_t idx = q-c;
+    p++;
+  }
 }
 #endif
